+++ CHANGELOG.md
@@ -22,6 +22,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Make REACT_APP_MIN_ZOOM_LEVEL optional
 - Improve use of MIN_ZOOM constant
 - Remove Search button
+- Enable cloud cover dependency based on collection
+- Remove time from date range
 
 ### Added
 
@@ -39,7 +41,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Minimum zoom level for search
 - Env variable REACT_APP_MIN_ZOOM_LEVEL
 - Add support for `asset_bidx` tiler configuration
-- Env variable REACT_APP_AWS_CF_TEMPLATE_URL
+- Env variable REACT_APP_CF_TEMPLATE_URL
 - Env variable REACT_APP_APP_NAME
 
 ## v0.2.0 - 2023-Jan-13